Struct brotlic::encode::BrotliEncoder

pub struct BrotliEncoder { /* private fields */ }

A reference to a brotli encoder.

This encoder contains internal state of the encoding process. This low-level wrapper intended to be used for people who are familiar with the C API. For higher level abstractions, see CompressorReader and CompressorWriter.

Implementations

impl BrotliEncoder

pub fn new() -> Self

Constructs a new brotli encoder instance.

Panics

Panics if the encoder fails to be allocated or initialized

pub fn is_finished(&self) -> bool

Checks if the encoder instance reached its final state.

pub fn compress( &mut self, input: &[u8], output: &mut [u8], op: BrotliOperation ) -> Result<EncodeResult, EncodeError>

Compresses input stream to output stream.

This is a low-level API, for higher level abstractions see CompressorReader or CompressorWriter. Returns the number of bytes that were read and written. Bytes are read from input, the number of bytes read is returned in the bytes_read field of the result. Bytes are written to output, the number of bytes written is returned in the bytes_written field of the result. The operation op specifies the intention behind this call, whether it is to simply process input, flush the input or finish the input. Care must be taken to not swap, reduce or extend the input stream while flushing or finishing. Additionally the operation should not change until all the input has been processed and all the output has been read from the internal buffer.

The internal workflow consists of three steps:

read from input into the internal buffer

compress input

write into output from the internal buffer

Whenever any of these tasks can’t move forward, control flow is returned to the caller. This is a wrapper around the BrotliEncoderCompressStream function of the C brotli API. For more information consult its documentation.

pub fn give_input( &mut self, input: &[u8], op: BrotliOperation ) -> Result<usize, EncodeError>

Convenience function to call method Self::compress with only input and no output.

pub fn flush(&mut self) -> Result<(), EncodeError>

Attempts the flush the encoding stream.

Actual flush is performed when all output has been successfully read. Use Self::has_output to verify that flushing completedNo other modifying operation should be queried before flushing has been finalized. When flush is complete, output data will be sufficient for a decoder to reproduce all given input. Calling this function might resulting in a worse compression ratio, because the encoder is forced to emit all output immediately.

pub fn finish(&mut self) -> Result<(), EncodeError>

Finalizes the encoding stream.

Actual finalization is performed when all output from the encoder has been successfully read. Use Self::is_finished to verify that the encoder is finished. Once this method has been called, no further input should be processed.

For more information, see BrotliEncoderOperation::BROTLI_OPERATION_FINISH

pub fn has_output(&self) -> bool

Checks if the encoder has more output.

pub unsafe fn take_output(&mut self) -> Option<&[u8]>

Checks if the encoder has more output and if so, returns a slice to its internal output buffer.

Each byte returned from the slice is considered “consumed” and must be used as it will not be returned again. Encoder output is not guaranteed to be contagious, which means that this function can return Some(&[u8]) multiple times. Only when the method returns None is when there is no more output available by the encoder.

Safety

For every consecutive call of this function, the previous slice becomes invalidated.

pub fn version() -> u32

Returns the version of the C brotli encoder library.