Struct brotlic::encode::CompressorReader

pub struct CompressorReader<R: BufRead> { /* private fields */ }

Wraps a reader and compresses its output.

The compression stream produced by brotli must be finished in order to be decompressed. This is done when the underlying reader reaches EOF. Therefore, CompressorReader<R> does not work with BufReads that return an infinite amount of data. When read returns zero on a non-zero buffer, the compression is considered finished.

Examples

Suppose the file test.txt contains uncompressed text. Letโs try to compress it:

use std::fs::File;
use std::io::Read;

use brotlic::DecompressorWriter;

let mut input = File::open("test.txt")?; // test.brotli is brotli compressed
let mut output = Vec::new();

input.read_to_end(&mut output)?;

println!("Compressed length: {}", output.len());

Implementations

impl<R: BufRead> CompressorReader<R>

pub fn new(inner: R) -> Self

Creates a new CompressorReader<R> with a newly created encoder.

Panics

Panics if the encoder fails to be allocated or initialized

pub fn with_encoder(encoder: BrotliEncoder, inner: R) -> Self

Creates a new CompressorReader<R> with a specified encoder.

Examples

use brotlic::{BrotliEncoderOptions, CompressorReader, Quality, WindowSize};

let encoder = BrotliEncoderOptions::new()
    .quality(Quality::new(6)?)
    .window_size(WindowSize::new(18)?)
    .build()?;

let underlying_source = [1, 2, 3, 4, 5];
let writer = CompressorReader::with_encoder(encoder, underlying_source.as_slice());

pub fn get_ref(&self) -> &R

Gets a reference to the underlying reader

pub fn get_mut(&mut self) -> &mut R

Gets a mutable reference to the underlying reader.

It is inadvisable to directly read from the underlying reader.

pub fn into_inner(self) -> Result<R, IntoInnerError<CompressorReader<R>>>

Unwraps this CompressorReader<R>, returning the underlying reader.

Errors

An Err will be returned if the compression stream has not been finished.

pub fn into_parts(self) -> (R, BrotliEncoder)

Disassembles this CompressorReader<R>, returning the underlying reader and encoder.

into_parts makes no attempt to validate that the compression stream finished and cannot fail.